requirements.txt
cryptography
pynput
pycryptodome
pillow
zxcvbn